refactor: AuctionFormPage로 분리 (#118)
